# Cache item info, reducing time of crafting filter `d:` to ~32%